Implement global.clusterVersion
We add global.clusterVersion variable (OCP major.minor) and pass it on.
Then we add a /values-<global.clusterVersion>-<clusterGroup>.yaml in the
helm parameter list. We wrap every inclusion of
/values-<ocpversion>-<clusterGroup>.yaml in an '{{ if
.global.clusterVersion }}' to avoid including anything in case the
variable is not defined or would end up having a default value which
would end up being undesirable.

This will allow us to have per-ocp version overrides and potentially
avoid having to use git branches more than needed.

Tested on MCG (w/ACM) and observed the correct global.clusterVersion
on both argo instances on hub and spoke. Also the new values includes
looked correct. I tested this as follows:

    diff --git a/charts/all/config-demo/templates/config-demo-cm.yaml b/charts/all/config-demo/templates/config-demo-cm.yaml
    index ac7fe991..795c235b 100644
    --- a/charts/all/config-demo/templates/config-demo-cm.yaml
    +++ b/charts/all/config-demo/templates/config-demo-cm.yaml
    @@ -5,6 +5,7 @@ metadata:
       labels:
	 app.kubernetes.io/instance: config-demo
     data:
    +  foo: {{ .Values.global.bandini }}
       "index.html": |-
	 <!DOCTYPE html>
	   <html lang="en">
    diff --git a/values-4.10-hub.yaml b/values-4.10-hub.yaml
    new file mode 100644
    index 00000000..350b40ec
    --- /dev/null
    +++ b/values-4.10-hub.yaml
    @@ -0,0 +1,3 @@
    +---
    +global:
    +  bandini: "4.10 is the best on the hub"
    diff --git a/values-4.9-group-one.yaml b/values-4.9-group-one.yaml
    new file mode 100644
    index 00000000..df704d75
    --- /dev/null
    +++ b/values-4.9.yaml
    @@ -0,0 +1,3 @@
    +---
    +global:
    +  bandini: "ZOMG I love 4.9 on the group-one spoke"

foo was "ZOMG I love 4.9 on the group-one spoke" on ocp 4.9.48 and "4.10
is the best on the hub" on
4.10.32

NOTE: This is for install-legacy *only* atm. A separate PR is needed to
add this as a top-level object in the operator.