Stop using OpenShiftControllerManager lookups
When applying the policy to install the cluster-wide argo on regional
clusters, we do some lookups() on the regional cluster in order to
pass the version values, domain names, etc.

To get the cluster version we were using the OpenShiftControllerManager
which is problematic because it does not exist on hyper-shift clusters.

Let's switch to use the ClusterVersion.status.history[0].version entry.

The only smaller caveat is that due to limitations in go templates +
sprig functions, we cannot really take the last version only when the
state is "Completed", but we simply take the last version.

This means that during a cluster upgrade on a regional cluster, we will
include values of the version the cluster is upgrading to, which is
less than ideal, but it should eventually converge in any case.

For reference the function that guarantees that the ordering of the
history status in ClusterVersion is preserved is here:
https://pkg.go.dev/github.com/openshift/api/config/v1#ClusterVersionStatus

Tested on Lester's cluster and on a local cluster of mine.

Co-Authored-By: Lester Claudio <claudiol@redhat.com>