Make HUB work when spokes point to in hub cluster gitea
With this change we push the hub CA to the spokes in both the
openshift-gitops, the namespaced argo namespace and in the imperative
namespace.

We do this so that the two argo instances are allowed to clone from an
eventual gitea repository on the hub. Same goes for the imperative
framework which might need to clone the code from the in-cluster gitea
service running on the hub.