Use lookups to determine the clusterPlatform
Right now we set global.clusterPlatform to whatever helm passes on, but
this is conceptually not correct:

        - name: global.clusterPlatform
          value: {{ $.Values.global.clusterPlatform }}

If the hub is on AWS and a spoke on Azure, the clusterPlatform on the
spoke needs to look it up locally, otherwise we can't do mixed
platforms.

Tested this on baremetal (where it was both None on hub and spoke) and
on AWS. The lookup works correctly so this should fix things properly.

Closes: #17