Use clustergroups on all policies
We don't want any policy that applies to just any cluster.
We need to make sure that every policy is tight to some
managedClusterGroup.

This way we avoid pushing stuff on clusters imported to ACM
that have nothing to do with the patterns.

We do this by creating a small template to support all the
existing ways to set labels.

Tested as follows:

  1. Deployed one hub and 4 spokes. Three group-one spokes and one
     group-two spokes via stock mcg -> All policies green and spokes
     correctly deployed everywhere

  2. Deployed one hub and 4 spokes via a private git repo.
     Three group-one spokes and one group-two spokes via stock mcg ->
     All policies green and spokes correctly deployed everywhere

Fixes: #16