Drop dependencies and make the configmaps non-optional
The dependencies were introduces so that the spoke cluster would have
all the right CAs from the hub, before creating the argo instance.

More often than not though these policy dependencies show up as yellow
for no clear reason (the dependencies show Compliant but the main policy
says it is not compliant due to the dependencies not being compliant)

So let's drop the dependencies and make the volume mounts for the argo
initcontainers non-optional. This way argo won't start until the
configmaps with the hub CAs are in place and we avoid this more complex
way of managing policies.